+++ packages/flow-form-builder/src/components/f-form-builder/fields/text.ts
@@ -3,6 +3,7 @@ import { FFormInputElements, FormBuilderField, FormBuilderTextInputField } from 
 import { Ref, ref } from "lit/directives/ref.js";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 import { getLabelLeftLayout, getSlots } from "../../../modules/helpers";
+import { getAriaLabel } from "../../../modules/utils";
 export default function (
 	name: string,
 	_field: FormBuilderField,
@@ -14,12 +15,13 @@ export default function (
 		<f-input
 			id=${ifDefined(field.id)}
 			class=${ifDefined(field.className)}
+			aria-label="${getAriaLabel(field) ?? name}"
 			name=${name}
 			.type=${field.type}
 			${ref(fieldRef)}
 			.placeholder=${field.placeholder}
 			.value=${value}
-			data-qa-element-id=${field.qaId || field.id}
+			data-qa-element-id=${ifDefined(field.qaId || field.id)}
 			icon-left=${ifDefined(field.iconLeft)}
 			icon-right=${ifDefined(field.iconRight)}
 			prefix=${ifDefined(field.prefix)}
@@ -31,13 +33,13 @@ export default function (
 			?disabled=${field.disabled ?? false}
 			?clear=${field.clear ?? true}
 			?read-only=${field.readonly ?? false}
-			@click=${ifDefined(field.onClick)}
-			@focus=${ifDefined(field.onFocus)}
-			@input=${ifDefined(field.onInput)}
-			@keypress=${ifDefined(field.onKeyPress)}
-			@keydown=${ifDefined(field.onKeyDown)}
-			@keyup=${ifDefined(field.onKeyUp)}
-			@mouseover=${ifDefined(field.onMouseOver)}
+			@click=${field.onClick}
+			@focus=${field.onFocus}
+			@input=${field.onInput}
+			@keypress=${field.onKeyPress}
+			@keydown=${field.onKeyDown}
+			@keyup=${field.onKeyUp}
+			@mouseover=${field.onMouseOver}
 			autofocus=${ifDefined(field.autofocus)}
 			autocomplete=${ifDefined(field.autocomplete)}
 		>