+++ packages/flow-core/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 # Change Log
 
+## [2.9.1] - 2024-03-10
+
+### Improvements
+
+- Externalise `lodash-es` to reduce bundle size.
+
+## [2.9.0] - 2024-03-01
+
+### Features
+
+- a11y support added.
+
 ## [2.8.7] - 2024-02-22
 
 ### Improvements